+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - tmp-fix: deleted lua queries #159
 - nvim-dev-icons colors fixed projekt0n/circles.nvim#6
 - make: tmux prefix fixed
+- tmux: set `bg2` as background in 'message-style' fixed #164
 
 ## [v0.0.4] - 11 Jan 2022